Add Neptune logging, fix import error
