+++ satflow/configs/logger/many_loggers.yaml
@@ -3,4 +3,4 @@
 
 defaults:
   - tensorboard.yaml
-  - csv.yaml
+  - neptune.yaml
