+++ satflow/configs/logger/neptune.yaml
@@ -1,0 +1,11 @@
+# https://neptune.ai
+
+neptune:
+  _target_: pytorch_lightning.loggers.neptune.NeptuneLogger
+  api_key: ${oc.env:NEPTUNE_API_TOKEN} # api key is loaded from environment variable
+  project_name: OpenClimateFix/forecasting-satellite-images
+  close_after_fit: True
+  offline_mode: False
+  experiment_name: null
+  experiment_id: null
+  prefix: ""
