+++ satflow/models/conv_lstm.py
@@ -7,7 +7,7 @@ import torch.nn as nn
 import numpy as np
 
 from satflow.models.base import register_model
-from satflow.models.utils import get_loss
+from satflow.models.losses import get_loss
 from satflow.models.layers.ConvLSTM import ConvLSTMCell
 import torchvision
