+++ satflow/models/losses.py
@@ -2,6 +2,7 @@ import numpy as np
 import torch
 import torch.nn as nn
 from pytorch_msssim import SSIM, MS_SSIM
+from torch.nn import functional as F
 
 
 class SSIMLoss(nn.Module):
@@ -116,3 +117,33 @@ class FocalLoss(nn.Module):
         else:
             loss = loss.sum()
         return loss
+
+
+def get_loss(loss: str = "mse", **kwargs) -> torch.nn.Module:
+    if isinstance(loss, torch.nn.Module):
+        return loss
+    assert loss in [
+        "mse",
+        "bce",
+        "binary_crossentropy",
+        "crossentropy",
+        "focal",
+        "ssim",
+        "ms_ssim",
+        "l1",
+    ]
+    if loss == "mse":
+        criterion = F.mse_loss
+    elif loss in ["bce", "binary_crossentropy", "crossentropy"]:
+        criterion = F.nll_loss
+    elif loss in ["focal"]:
+        criterion = FocalLoss()
+    elif loss in ["ssim"]:
+        criterion = SSIMLoss(data_range=1.0, size_average=True, **kwargs)
+    elif loss in ["ms_ssim"]:
+        criterion = MS_SSIMLoss(data_range=1.0, size_average=True, **kwargs)
+    elif loss in ["l1"]:
+        criterion = torch.nn.L1Loss()
+    else:
+        raise ValueError(f"loss {loss} not recognized")
+    return criterion
